# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Filkx API",
        "description": "API documentation for the Tech Filkx e-commerce platform",
        "contact": {
            "email": "admin@filkx.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://dev.api.tech.filkx.com",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/v1/wishlist": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "List the current user's wishlist",
                "operationId": "5c7490f8646ce567298c860f5c1e6ab8",
                "responses": {
                    "200": {
                        "description": "Wishlist items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/WishlistItemResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/wishlist/{product}": {
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Add a product to the wishlist",
                "operationId": "bc9a935679f7497ec1527419224a0aa1",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notify_on_drop": {
                                        "type": "boolean",
                                        "default": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product added to the wishlist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "price_at_add": {
                                            "type": "number",
                                            "format": "float",
                                            "nullable": true
                                        },
                                        "notify": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Remove a product from the wishlist",
                "operationId": "6dc795e21b581f62a4754522af254bd4",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product removed from the wishlist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/wishlist/{product}/notify": {
            "patch": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Toggle price-drop notifications for a wishlisted product",
                "operationId": "0fd460c58eb91ba9608ba2f725288e4f",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification preference toggled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "notify_on_drop": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "WishlistItemResource": {
                "title": "Wishlist Item Resource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "name": {
                        "description": "Localized product name, keyed by locale",
                        "type": "object"
                    },
                    "current_price": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "price_at_add": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "price_drop_pct": {
                        "description": "Percentage the price has dropped since it was added, if any",
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "notify_on_drop": {
                        "type": "boolean"
                    },
                    "added_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}